rename some classes for consistency, update coments as per PR review
