+++ src/json-rpc-api.ts
@@ -1,7 +1,7 @@
 import { JsonRpcRouter } from './lib/json-rpc-router.js';
 
 import { handleDwnProcessMessage } from './json-rpc-handlers/dwn/index.js';
-import { handleSubscriptionsClose } from './json-rpc-handlers/subscriptions/index.js';
+import { handleSubscriptionsClose } from './json-rpc-handlers/subscription/index.js';
 
 export const jsonRpcApi = new JsonRpcRouter();
